wk5 part 2, inc tests

+++ src/main.rs
@@ -14,7 +14,7 @@ use cli::Options;
 fn run_bft(opt: &Options) -> Result<(), Box<dyn std::error::Error>> {
     let program = BFProgram::from_file(&opt.file_name)?;
     // TODO don't like how the path's type was automatically specified above, but had to be manually specified below. Is there a way to avoid this?
-    let vm: VM<usize, &path::PathBuf> = VM::new(&program, 0, false);
+    let vm: VM<u8, &path::PathBuf> = VM::new(&program, 0, false);
     vm.interpret(&program);
 
     Ok(())
